chore: make clang-tidy happy

+++ libtransmission/rpcimpl.cc
@@ -1412,7 +1412,7 @@ char const* portTest(tr_session* session, tr_variant* args_in, tr_variant* /*arg
     if (tr_variantDictFindInt(args_in, TR_KEY_ip_protocol, &arg))
     {
         options.ip_proto = arg == 0 ? tr_web::FetchOptions::IPProtocol::V4 : tr_web::FetchOptions::IPProtocol::V6;
-        tr_variantDictAddInt(idle_data->args_out, TR_KEY_ip_protocol, arg != 0);
+        tr_variantDictAddInt(idle_data->args_out, TR_KEY_ip_protocol, static_cast<int64_t>(arg != 0));
     }
     session->fetch(std::move(options));
     return nullptr;
